==> 07_Seurat_Integration.Rmd <==
---
title: "241001_Seurat_Integration"
author: "Eileen Owens"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(dev = "png",
                      dpi = 300,
                      echo = TRUE,
                      cache = TRUE,
                      warning = FALSE,
                      keep_md = TRUE)
```

# Introduction
The purpose of this script is to integrate scRNA-seq data from all samples arising from the same location (thymus, lymp node), which can help match shared cell types and states across datasets, boost statistical power, and facilitate accurate comparative analysis across datasets. The integration procedure aims to return a single dimensional reduction that captures the shared sources of variance across multiple layers, so that cells in a similar biological state will cluster. The method returns a dimensional reduction which can be used for visualization and unsupervised clustering analysis.

# Loading data and packages
## Load necessary software packages
```{r software-load, results="hide", message=FALSE}
library(Seurat)
library(tidyverse)
library(clustree)
library(stringr)
library(DoubletFinder)
library(patchwork)
library(scales)
library(cowplot)
library(ggrepel)
library(colorspace)
library(DESeq2)
library(pheatmap)
library(RColorBrewer)
library(SeuratDisk)
library(SingleR)
library(viridis)
library(reshape)
library(lemon)
library(ggsankey)
library(msigdbr)
library(clusterProfiler)
library(slingshot)
library(ggpubr)
library(scRNAseq)
library(scuttle)
library(ape)
library(ggtree)
```

## Set working directory
```{r setwd}
setwd("C:/Users/edlarsen/Documents/240828_scRNAseq")
```

# Load filtered Seurat objects
```{r load-data}
filtered_ln <- readRDS(file = "C:/Users/edlarsen/Documents/240828_scRNAseq/seurat_merged_LN_filtered.RData")
filtered_thym <- readRDS(file = "C:/Users/edlarsen/Documents/240828_scRNAseq/seurat_merged_THYM_filtered.RData")
```

## Split layers prior to integration
```{r split-layers}
filtered_ln[["RNA"]] <- split(filtered_ln[["RNA"]], f = filtered_ln$orig.ident) # lymph node 
filtered_thym[["RNA"]] <- split(filtered_thym[["RNA"]], f = filtered_thym$orig.ident) # thymus
```

# Analyze data prior to integration

These parameters were selected based on recommendations in the Seurat documentation for performing integration: https://satijalab.org/seurat/articles/integration_introduction 

```{r integration1-LN, fig.width=14}
# run standard anlaysis workflow
## Lymph  node
integ_ln <- NormalizeData(filtered_ln)
integ_ln <- FindVariableFeatures(integ_ln)
integ_ln <- ScaleData(integ_ln)
integ_ln <- RunPCA(integ_ln)

# visualize
integ_ln <- FindNeighbors(integ_ln, dims = 1:30, reduction = "pca")
integ_ln <- FindClusters(integ_ln, resolution = 2, cluster.name = "unintegrated_LN_clusters")
integ_ln <- RunUMAP(integ_ln, dims = 1:30, reduction = "pca", reduction.name = "umap.unintegrated.LN")
DimPlot(integ_ln, reduction = "umap.unintegrated.LN", group.by = c("orig.ident", "seurat_clusters")) + # orig.ident is the sample ID
  plot_annotation(title = "Canine Lymph Node - Unintegrated", theme = theme(plot.title = element_text(hjust=0.5, size = 20)))
```

```{r integration1-THYM, fig.width=14}
# run standard analysis workflow
## Thymus
integ_thym <- NormalizeData(filtered_thym)
integ_thym <- FindVariableFeatures(integ_thym)
integ_thym <- ScaleData(integ_thym)
integ_thym <- RunPCA(integ_thym)

# visualize
integ_thym <- FindNeighbors(integ_thym, dims = 1:30, reduction = "pca")
integ_thym <- FindClusters(integ_thym, resolution = 2, cluster.name = "unintegrated_thym_clusters")
integ_thym <- RunUMAP(integ_thym, dims = 1:30, reduction = "pca", reduction.name = "umap.unintegrated.thym")
DimPlot(integ_thym, reduction = "umap.unintegrated.thym", group.by = c("orig.ident", "seurat_clusters")) + # orig.ident is the sample ID
  plot_annotation(title = "Canine Thymus - Unintegrated", theme = theme(plot.title = element_text(hjust=0.5, size = 20)))
```

# Perform integration
Canonical correlation analysis (CCA) identifies shared sources of variation between samples, but only identifies the greatest sources of variation in the data if it is shared or conserved across samples. CCA-based integration therefore enables integrative analysis when experimental conditions or disease states introduce very strong expression shifts, or when integrating datasets across modalities and species. However, CCA-based integration may also lead to overcorrection, especially when a large proportion of cells are non-overlapping across datasets.

RPCA-based integration runs faster, and represents a more conservative approach where cells in different biological states are less likely to 'align' after integration. RPCA is recommended when a substantial fraction of cells in one dataset have no matching type in the other, datasets originate from the same platform (i.e., multiple lanes of 10x genomics), or there are a large number of datasets or cells to integrate.

```{r integration2-LN}

#### Returns error: Error in UseMethod(generic = "Assays", object = object) : no applicable method for 'Assays' applied to an object of class "NULL"
integ_ln <- IntegrateLayers(object = integ_ln,
                            method = CCAIntegration,
                            orig.reduction = "pca",
                            new.reduction = "integrated.cca",
                            verbose = FALSE)
integ_ln[["RNA"]] <- JoinLayers(integ_ln[["RNA"]]) # rejoin layers after integration
```

```{r integration2-THYM}
integ_thym <- IntegrateLayers(object = integ_thym,
                            method = CCAIntegration,
                            orig.reduction = "pca", 
                            new.reduction = "integrated.cca", 
                            verbose = FALSE)
integ_thym[["RNA"]] <- JoinLayers(integ_thym[["RNA"]]) # rejoin layers after integration
```

## Visualize after integration
```{r postintegration-LN, fig.width=14}
## Lymph node 
integ_ln <- FindNeighbors(integ_ln, reduction = "integrated.cca", dims = 1:30)
integ_ln <- FindClusters(integ_ln, resolution = 1)
integ_ln <- RunUMAP(integ_ln, dims = 1:30, reduction = "integrated.cca")
DimPlot(integ_ln, reduction = "umap", group.by = c("orig.ident", "seurat_clusters")) +
  plot_annotation(title = "Canine Lymph Node - Integrated", theme = theme(plot.title = element_text(hjust=0.5, size = 20)))
```

```{r postintegration-THYM, fig.width=14}
## Thymus
integ_thym <- FindNeighbors(integ_thym, reduction = "integrated.cca", dims = 1:30)
integ_thym <- FindClusters(integ_thym, resolution = 1)
integ_thym <- RunUMAP(integ_thym, dims = 1:30, reduction = "integrated.cca")
DimPlot(integ_thym, reduction = "umap", group.by = c("orig.ident", "seurat_clusters")) +
  plot_annotation(title = "Canine Thymus - Integrated", theme = theme(plot.title = element_text(hjust=0.5, size = 20)))
```


## Save integrated Seurat objects
```{r save-object}
saveRDS(integ_ln, file="seurat_integrated_LN_filtered.RData")
saveRDS(integ_thym, file="seurat_integrated_THYM_filtered.RData")
```

# Citations
```{r citations}
sessionInfo()
citation()
```
