==> 13-Monocle3-PseudotimeTrajectoryAnalysis.Rmd <==
---
title: "Pseudotime Trajectory Analysis"
author: "Eileen Owens"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(dev = "png",
                      dpi = 300,
                      echo = TRUE,
                      cache = TRUE,
                      cache.lazy = FALSE,
                      warning = FALSE,
                      message = FALSE,
                      keep_md = TRUE)
```

# Introduction
The purpose of this script is to perform a pseudotime trajectory analysis on single-cell RNA-seq data from normal canine T-cell populations derived from the thymus and lymph node of healthy dogs.

# Acknowledgements
This script was adapted from the Monocle 3 documentation (https://cole-trapnell-lab.github.io/monocle3/docs/trajectories/) and "Single Cell RNA seq analysis - SEurat and Monocle3 pipeline" by Mahima Bose (https://rpubs.com/mahima_bose/Seurat_and_Monocle3_p).

# Software
```{r install, eval= FALSE, include = FALSE, echo=FALSE}
# First time: Install monocle3
## install Bioconductor dependencies
BiocManager::install(c('BiocGenerics', 'DelayedArray', 'DelayedMatrixStats',
                       'limma', 'lme4', 'S4Vectors', 'SingleCellExperiment',
                       'SummarizedExperiment', 'batchelor', 'HDF5Array',
                       'terra', 'ggrastr'))
# install monocle3
devtools::install_github('cole-trapnell-lab/monocle3')

# ensure 'promises' is version 1.3.2 or higher
packageVersion("promises")
#install_version("promises", >=1.3.2)

# install SeuratWrappers
remotes::install_github('satijalab/seurat-wrappers')
```

```{r software-load, results="hide", message=FALSE}
library(Seurat)
library(knitr)
library(tidyverse)
library(patchwork)
library(monocle3)
library(SeuratWrappers)
library(pheatmap)
library(stringr)
library(RColorBrewer)
```

# Data
Input: A Seurat object of T-cell subsets from normal canine lymph node and thymus.
```{r data-import, results='hide', message=FALSE}
setwd("C:/Users/edlarsen/Documents/240828_scRNAseq/Pseudotime")
seu <- readRDS(file = "C:/Users/edlarsen/Documents/240828_scRNAseq/T_Cells/IntegThymAndLN_Annotated.RData")
```

# Trajectory analysis with Monocle3
As cells move between biologic states, they undergo transcriptional reconfiguration, with some genes silenced and others newly activated. Trying to purify cells in these transient states to characterize them can be challenging. Monocle 3 utilizes an algorithm that learns the sequence of gene expression changes each cell must go through as it passes through dynamic biologic processes. Once it learns the overall trajectory of gene expression changes, it places each cell at its proper position in the trajectory. Monocle 3's differential expression analysis toolkit can then find genes regulated over the course of the trajectory.

If there are multiple outcomes for a process, a branched trajectory will be made, which correspond to cellular "decisions." The differential expression analysis toolkit is therefore also useful for identifying genes affected by and involved in making these decisions. 

### Convert Seurat object to celldata set object:
```{r seurat-to-cds}
cds <- as.cell_data_set(seu, assay = "RNA")
cds <- estimate_size_factors(cds)
fData(cds)$gene_short_name <- rownames(fData(cds))

## To view cell metadata:
# head(colData(cds))
## To view count data:
#head(counts(cds))
```

### Cluster data with Monocle 3
Performs unsupervised clustering of cells using Leiden community detection.
```{r cluster_cds1, fig.height=6, fig.width=9}
cds <- cluster_cells(cds, reduction_method = "UMAP")

plot_cells(cds, group_label_size = 5, show_trajectory_graph = FALSE) + plot_annotation("Monocle 3 clusters by Leiden community detection")

plot_cells(cds, color_cells_by = "IntegratedClusterIDs", group_label_size = 4, show_trajectory_graph = FALSE) + plot_annotation(title = "Original Seurat clusters") + theme(legend.position = "right")

```

### Add cluster names and plot clusters before trajectory
```{r cluster_cds2, fig.width=9, fig.height=6}
# create new column in colData(cds) an initialize it with values of partition(cds)
colData(cds)$assigned_cell_type <- as.character(clusters(cds))

# remap each cluster to cell type
colData(cds)$assigned_cell_type <- dplyr::recode(colData(cds)$assigned_cell_type,
                                                 "1" = "DP_Thymocytes_2",
                                                 "2" = "DP_Thymocytes_1",
                                                 "3" = "Naive_to_Activated_T",
                                                 "4" = "Naive_T_2",
                                                 "5" = "SP_Thym_and_Naive_T",
                                                 "6" = "Naive_T_1",
                                                 "7" = "Activated_T",
                                                 "8" = "SP_Thym",
                                                 "9" = "Proliferating_DP_Thym",
                                                 "10" = "CD8_NK",
                                                 "11" = "ETP_and_DN_Thym",
                                                 "12" = "Late_SP_Thym")

plot_cells(cds, 
           color_cells_by = "assigned_cell_type", 
           group_label_size = 4, 
           show_trajectory_graph = FALSE) + theme(legend.position = "right")
```

```{r eval=FALSE, include=FALSE, echo=FALSE}
# # Alternatively, clusters can be transferred from Seurat object. This chunk is set to eval=FALSE by default.
# 
# # assign partitions
# recreate.partitions <- c(rep(1, length(cds@colData@rownames))) # generates vector where the value 1 is repeated for the same length as the number of rows in cds@colData
# names(recreate.partitions) <- cds@colData@rownames # creates a named vector where each element corresponds to a row in cds@colData
# recreate.partitions <- as.factor(recreate.partitions) # converts vector into factor
# cds@clusters@listData[["UMAP"]][["partitions"]] <- recreate.partitions # assigns the value of recreate.partitions to a "partitions" element inside the "UMAP" structure within the listData slot of the clusters object in cds
# 
# # assign cluster info
# list.cluster <- seu@active.ident
# cds@clusters@listData[["UMAP"]][["clusters"]] <- list.cluster
# 
# # assign umap coordinates
# cds@int_colData@listData[["reducedDims"]]@listData[["UMAP"]] <- seu@reductions$umap@cell.embeddings
```

# Reduce dimensionality
```{r dim-reduce, fig.width = 10, fig.height = 8}
cds <- reduce_dimension(cds)
plot_cells(cds,
           label_groups_by_cluster = FALSE,
           group_label_size = 4,
           color_cells_by = "assigned_cell_type") + theme(legend.position = "right")

marker_genes1 <- c("CD4",
                   "GATA3",
                   "ZBTB7B",
                   "CD8A",
                   "RUNX3")

marker_genes2 <- c("CD34",
                   "KIT",
                   "RAG1",
                   "CCR9",
                   "MKI67")
  
marker_genes3 <- c("IL2RA",
                   "LTB",
                   "S1PR1",
                   "SELL",
                   "KLRB1")

plot_cells(cds,
           genes=marker_genes1,
           label_cell_groups=FALSE,
           show_trajectory_graph=FALSE)
plot_cells(cds,
           genes=marker_genes2,
           label_cell_groups=FALSE,
           show_trajectory_graph=FALSE)
plot_cells(cds,
           genes=marker_genes3,
           label_cell_groups=FALSE,
           show_trajectory_graph=FALSE)
```



## Learn trajectory
```{r trajectory-plot, results='hide', fig.width = 10, fig.height = 8}
cds <- cluster_cells(cds, reduction_method = "UMAP")
cds <- learn_graph(cds, use_partition = F)

plot_cells(cds,
           color_cells_by = "assigned_cell_type",
           label_groups_by_cluster = T,
           label_branch_points = T, 
           label_roots = T, 
           label_leaves = F,
           group_label_size = 5) + ggtitle("Trajectory analysis: Cell type annotations")

plot_cells(cds,
           label_branch_points = T, 
           label_roots = T, 
           label_leaves = F,
           group_label_size = 7) + ggtitle("Trajectory analysis: Monocle3 cluster numbers")
```

# Order cells in pseudotime
To place cells in order, Monocle 3 must be told where the "beginning" of the biologic process is, done by choosing regions on the graph to mark as "roots" of the trajectory.
```{r pseudotime-plot, fig.width = 10, fig.height = 8, message=FALSE, warning=FALSE}
cds <- order_cells(cds, reduction_method = "UMAP", root_cells = colnames(cds[, clusters(cds) == 10])) # use Monocle3 cluster numbers from figure above
plot_cells(cds,
           color_cells_by = "pseudotime",
           label_branch_points = T,
           graph_label_size=3,
           label_roots = F,
           label_leaves = F)
```

## Cells ordered by pseudotime
```{r ordered-by-pseudotime, fig.width = 10}
cds$monocle3_pseudotime <- pseudotime(cds)
data.pseudo <- as.data.frame(colData(cds))
ggplot(data.pseudo,
       aes(monocle3_pseudotime,
           reorder(assigned_cell_type, monocle3_pseudotime),
           fill = assigned_cell_type)) + geom_boxplot()
```

# Citations
```{r citations}
sessionInfo()
citation()
```
